shipit_code_coverage: Move chunk mapping code to a separate file and write tests for it
Also rename report_generators.py to zero_coverage.py.

Fixes #1031.
