+++ src/shipit_code_coverage/shipit_code_coverage/codecov.py
@@ -1,11 +1,7 @@
 # -*- coding: utf-8 -*-
-import concurrent.futures
-import json
 import os
 import shutil
-import sqlite3
 import tarfile
-from concurrent.futures import ThreadPoolExecutor
 from datetime import datetime
 from datetime import timedelta
 
@@ -17,14 +13,15 @@ from cli_common.log import get_logger
 from cli_common.taskcluster import get_service
 from cli_common.utils import ThreadPoolExecutorResult
 from cli_common.utils import retry
+from shipit_code_coverage import chunk_mapping
 from shipit_code_coverage import grcov
-from shipit_code_coverage import report_generators
 from shipit_code_coverage import taskcluster
 from shipit_code_coverage import uploader
 from shipit_code_coverage.artifacts import ArtifactsHandler
 from shipit_code_coverage.github import GitHubUtils
 from shipit_code_coverage.notifier import Notifier
 from shipit_code_coverage.secrets import secrets
+from shipit_code_coverage.zero_coverage import ZeroCov
 
 logger = get_logger(__name__)
 
@@ -122,53 +119,6 @@ class CodeCov(object):
 
             logger.info('Suite report generated', suite=suite)
 
-    def generate_chunk_mapping(self):
-        with ThreadPoolExecutor(max_workers=4) as executor:
-            futures = {}
-            for platform in ['linux', 'windows']:
-                for chunk in self.artifactsHandler.get_chunks():
-                    future = executor.submit(grcov.files_list, self.artifactsHandler.get(platform=platform, chunk=chunk), source_dir=self.repo_dir)
-                    futures[future] = (platform, chunk)
-
-            with sqlite3.connect('chunk_mapping.sqlite') as conn:
-                c = conn.cursor()
-                c.execute('CREATE TABLE file_to_chunk (path text, platform text, chunk text)')
-                c.execute('CREATE TABLE chunk_to_test (platform text, chunk text, path text)')
-
-                for future in concurrent.futures.as_completed(futures):
-                    (platform, chunk) = futures[future]
-                    files = future.result()
-                    c.executemany('INSERT INTO file_to_chunk VALUES (?,?,?)', ((f, platform, chunk) for f in files))
-
-                try:
-                    # Retrieve chunk -> tests mapping from ActiveData.
-                    r = requests.post('https://activedata.allizom.org/query', data=json.dumps({
-                        'from': 'unittest',
-                        'where': {'and': [
-                            {'eq': {'repo.branch.name': 'mozilla-central'}},
-                            {'eq': {'repo.changeset.id12': self.revision[:12]}},
-                            {'or': [
-                                {'prefix': {'run.key': 'test-linux64-ccov'}},
-                                {'prefix': {'run.key': 'test-windows10-64-ccov'}}
-                            ]}
-                        ]},
-                        'limit': 50000,
-                        'select': ['result.test', 'run.key']
-                    }))
-
-                    tests_data = r.json()['data']
-
-                    task_names = tests_data['run.key']
-                    test_iter = enumerate(tests_data['result.test'])
-                    chunk_test_iter = ((taskcluster.get_platform(task_names[i]), taskcluster.get_chunk(task_names[i]), test) for i, test in test_iter)
-                    c.executemany('INSERT INTO chunk_to_test VALUES (?,?,?)', chunk_test_iter)
-                except KeyError:
-                    # ActiveData is failing too often, so we need to ignore the error here.
-                    logger.error('Failed to retrieve chunk to tests mapping from ActiveData.')
-
-        with tarfile.open('chunk_mapping.tar.xz', 'w:xz') as tar:
-            tar.add('chunk_mapping.sqlite')
-
     def go(self):
         if self.from_pulse:
             commit_sha = self.githubUtils.get_commit(self.revision)
@@ -222,10 +172,10 @@ class CodeCov(object):
 
             self.generate_suite_reports()
 
-            zc = report_generators.ZeroCov(self.repo_dir)
-            zc.zero_coverage(self.artifactsHandler.get(), self.revision, self.github_revision)
+            zc = ZeroCov(self.repo_dir)
+            zc.generate(self.artifactsHandler.get(), self.revision, self.github_revision)
 
-            self.generate_chunk_mapping()
+            chunk_mapping.generate(self.repo_dir, self.revision, self.artifactsHandler)
 
             # Index the task in the TaskCluster index.
             self.index_service.insertTask(
